+++ NEWS.md
@@ -8,6 +8,7 @@ portupgrade 2.4.14
 
 * portupgrde -p: Fix new dependencies not getting installed. (#58)
 * Support security vulnerabilities as a build failure reason.
+* Fix ordering of build based on FETCH/PATCH/EXTRACT dependencies. [ports/177365]
 
 portupgrade 2.4.13 (released 2014-10-06)